+++ flask_admin/contrib/sqla/filters.py
@@ -3,14 +3,15 @@ import typing as t
 
 from sqlalchemy.sql import not_
 from sqlalchemy.sql import or_
-from sqlalchemy.sql.schema import Column
 
+from flask_admin._types import T_INSTRUMENTED_ATTRIBUTE
 from flask_admin._types import T_OPTIONS
-from flask_admin._types import T_SQLALCHEMY_QUERY
+from flask_admin._types import T_SQLALCHEMY_COLUMN
 from flask_admin._types import T_TRANSLATABLE
 from flask_admin._types import T_WIDGET_TYPE
 from flask_admin.babel import lazy_gettext
 from flask_admin.contrib.sqla import tools
+from flask_admin.contrib.sqla._types import T_SQLALCHEMY_QUERY
 from flask_admin.model import filters
 
 
@@ -21,7 +22,7 @@ class BaseSQLAFilter(filters.BaseFilter):
 
     def __init__(
         self,
-        column: Column,
+        column: T_SQLALCHEMY_COLUMN | T_INSTRUMENTED_ATTRIBUTE,
         name: str,
         options: T_OPTIONS = None,
         data_type: T_WIDGET_TYPE = None,
@@ -42,7 +43,9 @@ class BaseSQLAFilter(filters.BaseFilter):
 
         self.column = column
 
-    def get_column(self, alias: t.Any) -> Column:
+    def get_column(
+        self, alias: t.Any
+    ) -> T_SQLALCHEMY_COLUMN | T_INSTRUMENTED_ATTRIBUTE:
         return self.column if alias is None else getattr(alias, self.column.key)
 
     def apply(
@@ -130,7 +133,7 @@ class FilterEmpty(BaseSQLAFilter, filters.BaseBooleanFilter):
 class FilterInList(BaseSQLAFilter):
     def __init__(
         self,
-        column: Column,
+        column: T_SQLALCHEMY_COLUMN,
         name: str,
         options: T_OPTIONS = None,
         data_type: T_WIDGET_TYPE = None,
@@ -247,7 +250,7 @@ class DateSmallerFilter(FilterSmaller, filters.BaseDateFilter):
 class DateBetweenFilter(BaseSQLAFilter, filters.BaseDateBetweenFilter):
     def __init__(
         self,
-        column: Column,
+        column: T_SQLALCHEMY_COLUMN,
         name: str,
         options: T_OPTIONS = None,
         data_type: T_WIDGET_TYPE = None,
@@ -291,7 +294,7 @@ class DateTimeSmallerFilter(FilterSmaller, filters.BaseDateTimeFilter):
 class DateTimeBetweenFilter(BaseSQLAFilter, filters.BaseDateTimeBetweenFilter):
     def __init__(
         self,
-        column: Column,
+        column: T_SQLALCHEMY_COLUMN,
         name: str,
         options: T_OPTIONS = None,
         data_type: T_WIDGET_TYPE = None,
@@ -335,7 +338,7 @@ class TimeSmallerFilter(FilterSmaller, filters.BaseTimeFilter):
 class TimeBetweenFilter(BaseSQLAFilter, filters.BaseTimeBetweenFilter):
     def __init__(
         self,
-        column: Column,
+        column: T_SQLALCHEMY_COLUMN,
         name: str,
         options: T_OPTIONS = None,
         data_type: T_WIDGET_TYPE = None,
@@ -362,7 +365,11 @@ class TimeNotBetweenFilter(TimeBetweenFilter):
 
 class EnumEqualFilter(FilterEqual):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         self.enum_class = column.type.enum_class  # type: ignore[attr-defined]
         super().__init__(column, name, options, **kwargs)
@@ -375,7 +382,11 @@ class EnumEqualFilter(FilterEqual):
 
 class EnumFilterNotEqual(FilterNotEqual):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         self.enum_class = column.type.enum_class  # type: ignore[attr-defined]
         super().__init__(column, name, options, **kwargs)
@@ -388,7 +399,11 @@ class EnumFilterNotEqual(FilterNotEqual):
 
 class EnumFilterEmpty(FilterEmpty):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         self.enum_class = column.type.enum_class  # type: ignore[attr-defined]
         super().__init__(column, name, options, **kwargs)
@@ -396,7 +411,11 @@ class EnumFilterEmpty(FilterEmpty):
 
 class EnumFilterInList(FilterInList):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         self.enum_class = column.type.enum_class  # type: ignore[attr-defined]
         super().__init__(column, name, options, **kwargs)
@@ -413,7 +432,11 @@ class EnumFilterInList(FilterInList):
 
 class EnumFilterNotInList(FilterNotInList):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         self.enum_class = column.type.enum_class  # type: ignore[attr-defined]
         super().__init__(column, name, options, **kwargs)
@@ -430,7 +453,11 @@ class EnumFilterNotInList(FilterNotInList):
 
 class ChoiceTypeEqualFilter(FilterEqual):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         super().__init__(column, name, options, **kwargs)
 
@@ -440,13 +467,13 @@ class ChoiceTypeEqualFilter(FilterEqual):
         column = self.get_column(alias)
         choice_type = None
         # loop through choice 'values' to try and find an exact match
-        if isinstance(column.type.choices, enum.EnumMeta):  # type: ignore[attr-defined]
-            for choice in column.type.choices:  # type: enum.Enum  # type: ignore[attr-defined]
+        if isinstance(column.type.choices, enum.EnumMeta):  # type: ignore[union-attr]
+            for choice in column.type.choices:  # type: enum.Enum  # type: ignore[union-attr]
                 if choice.name == user_query:
                     choice_type = choice.value
                     break
         else:
-            for type, value in column.type.choices:  # type: ignore[attr-defined]
+            for type, value in column.type.choices:  # type: ignore[union-attr]
                 if value == user_query:
                     choice_type = type
                     break
@@ -458,7 +485,11 @@ class ChoiceTypeEqualFilter(FilterEqual):
 
 class ChoiceTypeNotEqualFilter(FilterNotEqual):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         super().__init__(column, name, options, **kwargs)
 
@@ -468,13 +499,13 @@ class ChoiceTypeNotEqualFilter(FilterNotEqual):
         column = self.get_column(alias)
         choice_type = None
         # loop through choice 'values' to try and find an exact match
-        if isinstance(column.type.choices, enum.EnumMeta):  # type: ignore[attr-defined]
-            for choice in column.type.choices:  # type: enum.Enum # type: ignore[attr-defined]
+        if isinstance(column.type.choices, enum.EnumMeta):  # type: ignore[union-attr]
+            for choice in column.type.choices:  # type: enum.Enum # type: ignore[union-attr]
                 if choice.name == user_query:
                     choice_type = choice.value
                     break
         else:
-            for type, value in column.type.choices:  # type: ignore[attr-defined]
+            for type, value in column.type.choices:  # type: ignore[union-attr]
                 if value == user_query:
                     choice_type = type
                     break
@@ -487,7 +518,11 @@ class ChoiceTypeNotEqualFilter(FilterNotEqual):
 
 class ChoiceTypeLikeFilter(FilterLike):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         super().__init__(column, name, options, **kwargs)
 
@@ -498,12 +533,12 @@ class ChoiceTypeLikeFilter(FilterLike):
         choice_types = []
         if user_query:
             # loop through choice 'values' looking for matches
-            if isinstance(column.type.choices, enum.EnumMeta):  # type: ignore[attr-defined]
-                for choice in column.type.choices:  # type: enum.Enum  # type: ignore[attr-defined]
+            if isinstance(column.type.choices, enum.EnumMeta):  # type: ignore[union-attr]
+                for choice in column.type.choices:  # type: enum.Enum  # type: ignore[union-attr]
                     if user_query.lower() in choice.name.lower():
                         choice_types.append(choice.value)
             else:
-                for type, value in column.type.choices:  # type: ignore[attr-defined]
+                for type, value in column.type.choices:  # type: ignore[union-attr]
                     if user_query.lower() in value.lower():
                         choice_types.append(type)
         if choice_types:
@@ -514,7 +549,11 @@ class ChoiceTypeLikeFilter(FilterLike):
 
 class ChoiceTypeNotLikeFilter(FilterNotLike):
     def __init__(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> None:
         super().__init__(column, name, options, **kwargs)
 
@@ -525,12 +564,12 @@ class ChoiceTypeNotLikeFilter(FilterNotLike):
         choice_types = []
         if user_query:
             # loop through choice 'values' looking for matches
-            if isinstance(column.type.choices, enum.EnumMeta):  # type: ignore[attr-defined]
-                for choice in column.type.choices:  # type: enum.Enum  # type: ignore[attr-defined]
+            if isinstance(column.type.choices, enum.EnumMeta):  # type: ignore[union-attr]
+                for choice in column.type.choices:  # type: enum.Enum  # type: ignore[union-attr]
                     if user_query.lower() in choice.name.lower():
                         choice_types.append(choice.value)
             else:
-                for type, value in column.type.choices:  # type: ignore[attr-defined]
+                for type, value in column.type.choices:  # type: ignore[union-attr]
                     if user_query.lower() in value.lower():
                         choice_types.append(type)
         if choice_types:
@@ -644,7 +683,7 @@ class FilterConverter(filters.BaseFilterConverter):
     arrow_type_filters = (DateTimeGreaterFilter, DateTimeSmallerFilter, FilterEmpty)
 
     def convert(
-        self, type_name: str, column: Column, name: str, **kwargs: t.Any
+        self, type_name: str, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter] | None:
         filter_name = type_name.lower()
 
@@ -672,19 +711,19 @@ class FilterConverter(filters.BaseFilterConverter):
         "IPAddressType",
     )
     def conv_string(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.strings]
 
     @filters.convert("UUIDType", "ColorType", "TimezoneType", "CurrencyType")
     def conv_string_keys(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.string_key_filters]
 
     @filters.convert("boolean", "tinyint")
     def conv_bool(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.bool_filters]
 
@@ -698,7 +737,7 @@ class FilterConverter(filters.BaseFilterConverter):
         "mediumint",
     )
     def conv_int(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.int_filters]
 
@@ -706,43 +745,47 @@ class FilterConverter(filters.BaseFilterConverter):
         "float", "real", "decimal", "numeric", "double_precision", "double"
     )
     def conv_float(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.float_filters]
 
     @filters.convert("date")
     def conv_date(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.date_filters]
 
     @filters.convert("datetime", "datetime2", "timestamp", "smalldatetime")
     def conv_datetime(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.datetime_filters]
 
     @filters.convert("time")
     def conv_time(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.time_filters]
 
     @filters.convert("ChoiceType")
     def conv_sqla_utils_choice(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.choice_type_filters]
 
     @filters.convert("ArrowType")
     def conv_sqla_utils_arrow(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.arrow_type_filters]
 
     @filters.convert("enum")
     def conv_enum(
-        self, column: Column, name: str, options: T_OPTIONS = None, **kwargs: t.Any
+        self,
+        column: T_SQLALCHEMY_COLUMN,
+        name: str,
+        options: T_OPTIONS = None,
+        **kwargs: t.Any,
     ) -> list[BaseSQLAFilter]:
         if not options:
             options = [(v, v) for v in column.type.enums]  # type: ignore[attr-defined]
@@ -751,6 +794,6 @@ class FilterConverter(filters.BaseFilterConverter):
 
     @filters.convert("uuid")
     def conv_uuid(
-        self, column: Column, name: str, **kwargs: t.Any
+        self, column: T_SQLALCHEMY_COLUMN, name: str, **kwargs: t.Any
     ) -> list[BaseSQLAFilter]:
         return [f(column, name, **kwargs) for f in self.uuid_filters]